Separate result frames from inTextSearch and add explicit clearing of right container
